Move `check_starts_not_past_stops()` to `utils.c`
